more rows

+++ usefulprograms/root.java
@@ -4,9 +4,9 @@ public class root {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		System.out.println("Enter a number: ");
-		double number = scan.nextInt();
+		double number = scan.nextDouble();
 		System.out.println("Enter the root: ");
-		double root = scan.nextInt();
+		double root = scan.nextDouble();
 		System.out.println("          _____________");
 		System.out.println((int)root + "        /     " + number + "        = " + root(number, root));
 		System.out.println("       \\/");
